+++ R/specreator.R
@@ -72,7 +72,7 @@
 #' provide a single spectrogram containing all elements for each song will be produce. Note that 
 #' the function assumes that each song has a unique label within a sound file. If \code{NULL} (default), spectrograms are produced for single selections.
 #' @param sel.labels Character string with the name of the column(s) for selection 
-#' labeling. Ignored if 'by.song' is \code{NULL}. Default is 'selec'. Set to \code{NULL} to remove labels.
+#' labeling. Default is 'selec'. Set to \code{NULL} to remove labels.
 #' @param title.labels Character string with the name(s) of the column(s) to use as title. Default is \code{NULL} (no title). Only sound file and song included if 'by.song' is provided.
 #' @param dest.path Character string containing the directory path where the image files will be saved.
 #' If \code{NULL} (default) then the folder containing the sound files will be used instead.